requirements.txt
amqp==5.0.9
anyjson==0.3.3
apturl==0.5.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
Brlapi==0.8.3
celery==5.2.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.0.6
django-celery-beat==2.4.0
django-db-logger==0.1.12
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.4
duplicity==0.8.21
ephem==4.1.3
fasteners==0.14.1
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.5.0
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
packaging==22.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.2.0
prompt-toolkit==3.0.28
protobuf==3.12.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.4.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.2
python-apt==2.3.0+ubuntu2.1
python-crontab==2.6.0
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-memcached==1.58
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.4.0
reportlab==3.6.8
requests==2.25.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.1
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==3.0.3
systemd-python==234
typing_extensions==4.4.0
tzdata==2022.1
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.5
xdg==5
xkit==0.0.0
zipp==1.0.0
